requirements.txt
pyOpenSSL==25.0.0
cryptography==44.0.0
